prevent stage/configure product/director if apply-change is running
If a product is staged/configured while apply-change is running,
the product will be unstaged/unconfigured after the apply-change
finishes. This is an unexpected behavior to the users, so we want
to prevent stage/configure product/director if there is a running
installation.

The change in this commit will fail fast if the cli user tries to
stage/configure if apply-change is running.

The prevention logic will eventually lives in Ops Manager as well.

[#161630968]